Cleaning up feature items
