+++ packages/web/src/components/features/index.ts
@@ -1,0 +1,2 @@
+export * from './Features'
+export * from './FeatureItem'
