+++ packages/web/src/styles/components/Features.module.scss
@@ -18,3 +18,7 @@
         gap: 2.5rem;
     }
 };
+
+.imageWrapper {
+    max-width: 100%;
+}
