+++ packages/web/src/styles/components/layout/RichTextWrapper.module.scss
@@ -1,6 +1,6 @@
 .wrapper {
     grid-column: 2;
-    padding: 0 2.5rem;
+    padding: 2.5rem;
 
     @media (max-width: 768px) {
         padding: 0 1rem;
